site/requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
cffi==1.11.5
cryptography==2.3.1
Django==2.1.1
djangorestframework==3.8.2
enum34==1.1.6
escpos==1.5
idna==2.7
ipaddress==1.0.22
paramiko==2.4.1
Pillow==5.2.0
pkg-resources==0.0.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyasn1==0.4.4
pycparser==2.18
PyNaCl==1.2.1
PyQRCode==1.2.1
pyserial==3.4
pytz==2018.5
pyusb==1.0.2
qrcode==6.0
six==1.11.0
